% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bqr.R
\name{bqr}
\alias{bqr}
\title{Bayesian quantile regression}
\usage{
bqr(x, y, tau, knots = 20, degree = 3, diff = 2, burnin = 1000,
  step = 10, iter = 1000, n0 = 100, s0 = 1, a = 0.001, b = 0.001,
  tau20 = 10)
}
\description{
Bayesian quantile regression
}
