% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotbarchi}
\alias{plotbarchi}
\title{Plot barchi}
\usage{
plotbarchi(fit, xlab = "t", n = 100)
}
\arguments{
\item{fit}{tailind output}

\item{xlab}{x-axis label}

\item{n}{Number of points for smooth line}
}
\description{
Plot barchi
}
