% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotchi}
\alias{plotchi}
\title{Plot chi}
\usage{
plotchi(fit, xlab = "t", n = 100, knots = 30, degree = 3)
}
\arguments{
\item{fit}{taildep output}

\item{xlab}{x-axis label}

\item{n}{Number of points for smooth line}
}
\description{
Plots a taildep ouput.
}
