% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCopula.R
\name{plotCopulaBarChi}
\alias{plotCopulaBarChi}
\title{Plot copula barchi}
\usage{
plotCopulaBarChi(copula, u = seq(0.9, 0.9999, length.out = 100), ylim = c(0,
  1), ...)
}
\description{
Plot copula barchi
}
\examples{
blevd <- function(u,v, ... ) exp(-( (-log(u))^(1/alpha)+(-log(v))^(1/alpha))^alpha )
alpha <- c(0.1,0.3,0.5,0.7,0.9)
plotCopulaBarChi(blevd, u=seq(0.9,0.9999, length.out = 100),
                 alpha=alpha  )
points(1, 1, col=2)
}
