% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCopula.R
\name{plotCopulaChi}
\alias{plotCopulaChi}
\title{Plot copula chi(u)}
\usage{
plotCopulaChi(copula, u = seq(0.9, 0.9999, length.out = 100), ylim = c(0,
  1), ...)
}
\description{
Plot copula chi(u)
}
\examples{
blevd <- function(u,v, ... ) exp(-( (-log(u))^(1/alpha)+(-log(v))^(1/alpha))^alpha )
alpha <- c(0.1,0.3,0.5,0.7,0.9)
plotCopulaChi(blevd, u=seq(0.9,0.9999, length.out = 100),
              alpha=alpha  )
points(rep(1, length(alpha)), 2-2^alpha, col=2)
}
