% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMeasures}
\alias{plotMeasures}
\title{Plot both measures chi_t and barchi_t}
\usage{
plotMeasures(fitdep, fitind, xlab = "t", ylim = c(-1, 1), n = 100,
  target = NULL)
}
\description{
Plot both measures chi_t and barchi_t
}
