% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindexreglogit.R
\name{tailindexreglogit}
\alias{tailindexreglogit}
\title{Tail index regression (logit)}
\usage{
tailindexreglogit(xy, t, threshold = 0.95, k = 30, l = 3, d = 1,
  atau2 = 0.001, btau2 = 0.001, step = 20, burnin = 2000, iter = 4000,
  lambda = 1, maxiter = 500, link = "logit", control.bqr = list(burnin =
  1000, step = 10, iter = 1000), bqr = FALSE)
}
\arguments{
\item{xy}{Matrix of data n rows and two columns.}

\item{t}{Time values. Vector of n.}

\item{threshold}{For quantile function. Default 0.95.}

\item{k}{Number of knots. Default 30.}

\item{l}{Degree of splines. Default 3.}

\item{d}{Diference between betas_k. Default 1.}

\item{atau2}{Hyperparameter a for tau2. Default 0.001}

\item{btau2}{Hyperparameter b for tau2. Default 0.001}

\item{step}{Thinning. Default 20}

\item{burnin}{First iterations discarted. Default 2000}

\item{iter}{Number of final iterations. Default 4000}

\item{lambda}{Penalization values. Default 1}

\item{maxiter}{Maximum of iterations for IWLS algorithm. Default 500}

\item{link}{String "logit" or "log"}
}
\value{
A list
}
\description{
Posterior samples of barchi(t).
}
