% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindexreglogitRint.R
\name{tailindexreglogit}
\alias{tailindexreglogit}
\title{Tail index regression (logit)}
\usage{
tailindexreglogit(xy, t, threshold = 0.95, k = 30, l = 3, d = 1,
  atau2 = 0.001, btau2 = 0.001, Vtau2 = NULL, asigma2 = 0.001,
  bsigma2 = 0.001, anu = 0.1, bnu = 0.1, varnup = 0.01, step = 10,
  burnin = 2000, iter = 4000, lambda = 1, maxiter = 500,
  link = "logit", theta0 = NULL, control.bqr = NULL)
}
\arguments{
\item{xy}{Matrix of data n rows and two columns.}

\item{t}{Time values. Vector of n.}

\item{threshold}{For quantile function. Default 0.95.}

\item{k}{Number of knots. Default 30.}

\item{l}{Degree of splines. Default 3.}

\item{d}{Diference between betas_k. Default 1.}

\item{atau2}{Hyperparameter a for tau2. Default 0.001}

\item{btau2}{Hyperparameter b for tau2. Default 0.001}

\item{Vtau2}{Variance of tau2 samples. Default NULL}

\item{asigma2}{Hyperparameter a for sigma2. Default 0.001}

\item{bsigma2}{Hyperparameter b for sigma2. Default 0.001}

\item{anu}{Hyperparameter scale parameter. Default 0.1}

\item{bnu}{Hyperparameter scale parameter. Default 0.1}

\item{varnup}{Proposal parameter. Default 0.01}

\item{step}{Thinning. Default 20}

\item{burnin}{First iterations discarted. Default 2000}

\item{iter}{Number of final iterations. Default 4000}

\item{lambda}{Penalization values. Default 1}

\item{maxiter}{Maximum of iterations for IWLS algorithm. Default 500}

\item{link}{Character. Default "logit" or "exp" only.}

\item{theta0}{Initial constant value. Default NULL.}
}
\value{
A list
}
\description{
Posterior samples of barchi(t).
}
